requirements.txt
mysql-connector-python
transformers
beautifulsoup4
requests
selenium
webdriver-manager